// src/components/atoms/General/ErrorPage/ErrorPage.enum.ts
export enum ERROR_PAGE_ELEMENTS {
  CONTAINER = "error-page-container",
  ICON = "error-page-icon",
  TITLE = "error-page-title",
  DESCRIPTION = "error-page-description",
  ERROR_CODE = "error-page-error-code",
  HOME_BUTTON = "error-page-home-button",
  RETRY_BUTTON = "error-page-retry-button",
  FOOTER = "error-page-footer",
}